+++ borax/__init__.py
@@ -1,4 +1,4 @@
 # coding=utf8
 
-__version__ = '3.5.1'
+__version__ = '3.5.2'
 __author__ = 'kinegratii'