+++ borax/htmls.py
@@ -1,6 +1,7 @@
 # coding=utf8
 
 import html
+from typing import Dict, List, Optional
 
 
 def _escape(s):
@@ -42,12 +43,35 @@ def html_params(**kwargs) -> str:
             params.append(k)
         elif v is False:
             pass
+        elif isinstance(v, dict):
+            vs = ''.join(['{}:{};'.format(k, v) for k, v in v.items()])
+            params.append('%s="%s"' % (str(k), vs))
+        elif isinstance(v, list):
+            vs = ' '.join(map(str, v))
+            params.append('%s="%s"' % (str(k), vs))
         else:
             params.append('%s="%s"' % (str(k), html.escape(str(v), quote=True)))
     return ' '.join(params)
 
 
-def html_tag(tag_name, content=None, **kwargs):
-    if content:
-        return HTMLString('<{0} {1}>{2}</{0}>'.format(tag_name, html_params(**kwargs), content))
-    return HTMLString('<{0} {1}>'.format(tag_name, html_params(**kwargs)))
+SINGLE_TAGS = ('br', 'hr', 'img', 'input', 'param', 'meta', 'link')
+
+
+def html_tag(tag_name: str, content: str = None,
+             *, id_: str = None, style: Dict = None, class_: Optional[List] = None, **kwargs) -> HTMLString:
+    """生成元素的html字符串"""
+    kw = {}
+    if id_:
+        kw['id_'] = id_
+    style = style or {}
+    if style:
+        kw['style'] = style
+    class_ = class_ or []
+    if class_:
+        kw['class_'] = class_
+    kw.update(kwargs)
+    if tag_name in SINGLE_TAGS:
+        return HTMLString('<{0} {1}>'.format(tag_name, html_params(**kw)))
+    else:
+        content = content or ''
+        return HTMLString('<{0} {1}>{2}</{0}>'.format(tag_name, html_params(**kw), content))